Moved MainWidget into display
added some environment variable that may help with system memory usage
